Modularize transformer HF benchmark
