+++ modeling/transformers/src/tilegym_hf_bench/__init__.py
@@ -1,0 +1,5 @@
+# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+#
+# SPDX-License-Identifier: MIT
+
+"""TileGym Hugging Face inference benchmark utilities."""
