+++ modeling/transformers/src/tilegym_hf_bench/__main__.py
@@ -1,0 +1,8 @@
+# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+#
+# SPDX-License-Identifier: MIT
+
+from tilegym_hf_bench._cli import main
+
+if __name__ == "__main__":
+    main()
