+++ modeling/transformers/src/tilegym_hf_bench/kernel_filters/__init__.py
@@ -1,0 +1,7 @@
+# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+#
+# SPDX-License-Identifier: MIT
+
+from tilegym_hf_bench.kernel_filters.filters import KernelFilter
+
+__all__ = ["KernelFilter"]
