+++ skills/tilegym-monkey-patch-kernels-to-transformers/references/auto-kernelize.md
@@ -6,17 +6,19 @@ Work with user to prepare experiment environment:
 1. Check Git branch status. The previous commit should only contain monkey-patching existing TileGym OPs to the target transformers model. No other unstaged/uncommitted modifications
 2. Check GPU available and UUID match; Docker container has been built
 3. Study code relating to the target transformers model:
-   - modeling/transformers/bench_<submodule_name>.sh: end-to-end benchmark entrance, run PyTorch baseline perf, cuTile kernelize perf, cuTile kernel coverage
+   - modeling/transformers/scripts/benchmark_hf_model.sh: end-to-end benchmark entrance, run PyTorch baseline perf, cuTile kernelize perf, cuTile kernel coverage
    - src/tilegym/transformers/<submodule_name>/modeling_<submodule_name>.py: Target model specific OP adapters and wrappers
    - src/tilegym/transformers/<submodule_name>/kernel_definitions/*.json and kernel_solutions/*.json: Existing reusable kernel inventory
    - src/tilegym/transformers/<submodule_name>/kernels/*.py: Dedicated reusable transformer-local kernel implementations
    - @src/tilegym/transformers/monkey_patch.py: Study apply_tilegym_kernel_to_<submodule_name> to understand how to kernelize
-   - @modeling/transformers/infer.py: End-to-end benchmark and kernel coverage script
+   - @modeling/transformers/src/tilegym_hf_bench/_cli.py: End-to-end benchmark and kernel coverage CLI
+   - @modeling/transformers/src/tilegym_hf_bench/tilegym_patch.py: model-id dispatch to TileGym monkey-patch functions
+   - @modeling/transformers/src/tilegym_hf_bench/kernel_filters/tilegym_kernel_prefixes.yaml: cuTile kernel coverage filter prefixes
 4. Create sandbox/<submodule_name>_results.md to track progress. The first run will write a baseline
 5. Confirm and go: Once you get confirmation, kick off the experimentation
 
 ## Experimentation
-Every experiment run on a NVIDIA GPU that [TileIR supported](https://docs.nvidia.com/cuda/tile-ir/latest/sections/stability.html#supported-architectures) (currently Ampere, Ada, and Blackwell architectures). Each experiment should be enforced to finish in 15 minutes. Every command should be executed within the experiment Docker container. `cd` to @modeling/transformers/ first, then `bash bench_<submodule_name>.sh` to launch one experiment.
+Every experiment must run on an NVIDIA GPU [supported by TileIR](https://docs.nvidia.com/cuda/tile-ir/latest/sections/stability.html#supported-architectures) (currently Ampere, Ada, and Blackwell). Each experiment should be enforced to finish in 15 minutes. Every command should be executed within the experiment Docker container. `cd` to @modeling/transformers/ first, then `bash scripts/benchmark_hf_model.sh --model-key <submodule_name>` to launch one experiment.
 
 Reusable generated kernels must follow the kernel inventory schema linked from `SKILL.md`: start each experiment with a draft FlashInfer Definition, keep verified kernels in dedicated `kernels/<kernel_name>.py` files, and record matching Definition and Solution JSON metadata. The Definition `reference` must begin with `# Source:` comment(s) pointing to precise upstream `transformers` or Hugging Face remote-code regions.
 
@@ -30,17 +32,16 @@ Reusable generated kernels must follow the kernel inventory schema linked from `
 - @src/tilegym/transformers/<submodule_name>/kernel_definitions/<kernel_name>.json: FlashInfer Definition metadata for kept reusable kernels
 - @src/tilegym/transformers/<submodule_name>/kernel_solutions/<kernel_name>.json: FlashInfer Solution metadata for kept reusable kernels
 - @src/tilegym/transformers/monkey_patch.py: Only change the `apply_tilegym_kernel_to_<submodule name>` function
-- @modeling/transformers/infer.py: Only change
-  * `apply_tilegym_kernel_to_<submodule name>` arguments
-  * `KernelFilter.kernel_names_prefix` for new cuTile kernels
-- @modeling/transformers/bench_<submodule_name>.sh: Optionally comment out line 30-38 to skip PyTorch throughput to accelerate experiment iterations. If so, ensure to restore on each experiment end
+- @modeling/transformers/src/tilegym_hf_bench/tilegym_patch.py: Only change `apply_tilegym_kernel_to_<submodule name>` dispatch arguments
+- @modeling/transformers/src/tilegym_hf_bench/kernel_filters/tilegym_kernel_prefixes.yaml: Only add kernel name substrings for new cuTile kernels after checking current nsys names
+- @modeling/transformers/scripts/benchmark_hf_model.sh: Optionally run with `--skip-baseline` to accelerate experiment iterations. Restore full baseline + cuTile + coverage at each experiment end
 - @sandbox/: Feel free to add new files or modify files created by you, but don't check to git
 
 ### What you can NOT change
 - Anything not listed above
 
 ### What to expect from experiment outputs
-`bench_<submodule_name>.sh` prints ~300 lines of plain text. Use this command to grep core metrics: `grep -E "Average throughput|cuTile Kernel Coverage \(GPU Time\)" <output_file>`. Example output:
+`scripts/benchmark_hf_model.sh --model-key <submodule_name>` prints ~300 lines of plain text. Use this command to grep core metrics: `grep -E "Average throughput|cuTile Kernel Coverage \(GPU Time\)" <output_file>`. Example output:
 
 ```text
 Average throughput: 25.93 ± 3.20 tokens/sec
@@ -69,7 +70,7 @@ Example content:
 Create the tabular header if the file was empty. Append one line for currently experiment.
 
 ### The baseline
-The first experiment will not change any code and simply run bench_<submodule_name>.sh. Results will list at first row as baseline.
+The first experiment will not change any code and simply run `scripts/benchmark_hf_model.sh --model-key <submodule_name>`. Results will list at first row as baseline.
 
 ## The experiment loop
 Core methodology is to create new cuTile kernels to replace uncovered PyTorch code while keeping performant and correctness. Try one piece of code at a time, and have clean experiment records.
